+++ src/group-26x/java/com/pvptweaks/util/ShieldSampleStack.java
@@ -1,0 +1,8 @@
+package com.pvptweaks.util;
+
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+
+public class ShieldSampleStack {
+    public static final ItemStack INSTANCE = new ItemStack(Items.SHIELD);
+}